+++ src/theme/augmentation.ts
@@ -6,6 +6,8 @@ import type { WizardStepperProps } from '../Wizard/WizardStepper'
 import type { WizardNavProps } from '../Wizard/WizardNav'
 import type { ReadOnlyFieldProps } from '../fields/ReadOnlyField'
 import type { NumberFieldControlProps } from '../fields/NumberField/NumberFieldControl'
+import type { FormProps } from '../Form'
+import type { FormSectionProps } from '../FormSection'
 
 declare module '@mui/material/styles' {
   interface ComponentsPropsList {
@@ -16,6 +18,9 @@ declare module '@mui/material/styles' {
     EzWizardNav: Partial<WizardNavProps>
     EzReadOnlyField: Partial<ReadOnlyFieldProps>
     EzNumberField: Partial<NumberFieldControlProps>
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    EzForm: Partial<FormProps<any, any>>
+    EzFormSection: Partial<FormSectionProps>
   }
 
   interface ComponentNameToClassKey {
@@ -26,6 +31,8 @@ declare module '@mui/material/styles' {
     EzWizardNav: 'root' | 'prev' | 'next' | 'submit'
     EzReadOnlyField: 'root' | 'header' | 'label' | 'value' | 'edit'
     EzNumberField: 'root' | 'steppers' | 'increment' | 'decrement'
+    EzForm: 'root' | 'title' | 'description'
+    EzFormSection: 'root' | 'legend' | 'description' | 'content'
   }
 
   interface Components<Theme = unknown> {
@@ -56,6 +63,14 @@ declare module '@mui/material/styles' {
     EzNumberField?: {
       defaultProps?: ComponentsProps['EzNumberField']
       styleOverrides?: ComponentsOverrides<Theme>['EzNumberField']
+    }
+    EzForm?: {
+      defaultProps?: ComponentsProps['EzForm']
+      styleOverrides?: ComponentsOverrides<Theme>['EzForm']
+    }
+    EzFormSection?: {
+      defaultProps?: ComponentsProps['EzFormSection']
+      styleOverrides?: ComponentsOverrides<Theme>['EzFormSection']
     }
   }
 }